# src/SluiceBase.Api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SluiceBase.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/server": {
      "get": {
        "tags": [
          "ServerEndpoints"
        ],
        "operationId": "ListServers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListServersResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ServerEndpoints"
        ],
        "operationId": "CreateServer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/server/{id}": {
      "put": {
        "tags": [
          "ServerEndpoints"
        ],
        "operationId": "UpdateServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ServerEndpoints"
        ],
        "operationId": "DeleteServer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionMode": {
        "enum": [
          "Standard",
          "Srv",
          null
        ]
      },
...
      "CreateServerRequest": {
        "required": [
          "name",
          "kind",
          "host",
          "port"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "connectionMode": {
            "$ref": "#/components/schemas/ConnectionMode"
          },
          "authSource": {
            "type": [
              "null",
              "string"
            ]
          },
          "replicaSet": {
            "type": [
              "null",
              "string"
            ]
          },
          "useTls": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "CredentialId": {
        "type": "string",
        "format": "uuid"
      },
      "CredentialResponse": {
        "required": [
          "id",
          "label",
          "username",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/CredentialId"
          },
          "label": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DatabaseId": {
        "type": "string",
        "format": "uuid"
      },
      "DatabaseResponse": {
        "required": [
          "id",
          "displayName",
          "databaseName",
          "readCredentialId",
          "writeCredentialId",
          "canWrite",
          "isDisabled",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/DatabaseId"
          },
          "displayName": {
            "type": "string"
          },
          "databaseName": {
            "type": "string"
          },
          "readCredentialId": {
            "$ref": "#/components/schemas/CredentialId"
          },
          "writeCredentialId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CredentialId"
              }
            ]
          },
          "canWrite": {
            "type": "boolean"
          },
          "isDisabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ListServersResponse": {
        "required": [
          "servers"
        ],
        "type": "object",
        "properties": {
          "servers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerResponse"
            }
          }
        }
      },
...
      "ServerId": {
        "type": "string",
        "format": "uuid"
      },
      "ServerResponse": {
        "required": [
          "id",
          "name",
          "kind",
          "host",
          "port",
          "connectionMode",
          "authSource",
          "replicaSet",
          "useTls",
          "isDisabled",
          "credentials",
          "databases",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ServerId"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "connectionMode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConnectionMode"
              }
            ]
          },
          "authSource": {
            "type": [
              "null",
              "string"
            ]
          },
          "replicaSet": {
            "type": [
              "null",
              "string"
            ]
          },
          "useTls": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "isDisabled": {
            "type": "boolean"
          },
          "credentials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CredentialResponse"
            }
          },
          "databases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatabaseResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateServerRequest": {
        "required": [
          "name",
          "host",
          "port",
          "kind"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "kind": {
            "type": "string"
          },
          "isDisabled": {
            "type": "boolean",
            "default": false
          },
          "connectionMode": {
            "$ref": "#/components/schemas/ConnectionMode"
          },
          "authSource": {
            "type": [
              "null",
              "string"
            ]
          },
          "replicaSet": {
            "type": [
              "null",
              "string"
            ]
          },
          "useTls": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    }
  },
...
}